chore: Refactor dynatrace client into its own authentication sub-folder
