+++ .vscode/mcp.json
@@ -1,0 +1,10 @@
+{
+  "servers": {
+    "my-dynatrace-mcp-server": {
+      "command": "node",
+      "cwd": "${workspaceFolder}",
+      "args": ["${workspaceFolder}/dist/index.js"],
+      "envFile": "${workspaceFolder}/.env"
+    }
+  }
+}
